requirements.txt
accelerate==0.22.0
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==3.2.1
Argumental==1.13
arrow==1.3.0
async-timeout==4.0.2
attrs==23.1.0
bardapi==0.1.11
Baubles==1.6
bitsandbytes==0.41.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.12.0
dill==0.3.6
dotmap==1.3.30
et-xmlfile==1.1.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gensim==4.3.1
GoldenChild==1.16
google-ai-generativelanguage==0.2.0
google-api-core==2.11.0
google-auth==2.19.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.0
greenlet==3.0.2
grpcio==1.54.2
grpcio-status==1.54.2
huggingface-hub==0.17.1
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
jsonpath==0.82.2
jsonpickle==3.0.2
kiwisolver==1.4.4
KnowledgeBase==1.12
langchain==0.0.181
Levenshtein==0.21.1
lxml==5.0.1
marshmallow==3.20.1
matplotlib==3.7.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
neo4j-driver==5.9.0
nltk==3.8.1
numexpr==2.8.8
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
Perdy==1.11
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==1.10.13
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==1.1.0
python-Levenshtein==0.21.1
pytz==2023.3
PyXB==1.2.6
PyYAML==6.0
rank-bm25==0.2.2
rapidfuzz==3.1.1
regex==2023.5.5
requests==2.31.0
responses==0.18.0
rsa==4.9
safetensors==0.3.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
SQLAlchemy==2.0.23
Swapsies==1.51
tenacity==8.2.3
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.65.0
transformers==4.33.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.6.1
tzdata==2023.3
underscore.py==0.1.6
urllib3==1.26.16
vobject==0.9.6.1
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
